refactor(SideNav): extract useResizable hook from XDSSideNav
Move all resize state, pointer event listeners, and drag logic into
a standalone useResizable hook. The hook only sets up event listeners
when resizable is enabled, and returns handle props + width state.
